[Server] / #44 / fix: sequelize 위도, 경도 관련 스키마 구조 변경

+++ server/migrations/20220103053012-create-store.js
@@ -16,11 +16,11 @@ module.exports = {
       },
       latitude: {
         allowNull: false,
-        type: Sequelize.DECIMAL
+        type: Sequelize.DOUBLE
       },
       longitude: {
         allowNull: false,
-        type: Sequelize.DECIMAL
+        type: Sequelize.DOUBLE
       }
     });
   },
